requirements.txt
imageio
hydra-core==1.2.0
omegaconf==2.3.0
tqdm
pytorch_lightning==1.5.9
torch==1.10.1
torchvision==0.11.2
einops==0.6.0
basicsr==1.4.2
transformers==4.27.3
opencv-python
gradio
kornia
timm
numpy==1.24.3
safetensors
open-clip-torch
scikit-learn

# --extra-index-url https://download.pytorch.org/whl/cu113
# torch==1.12.1+cu113 
# torchvision==0.13.1+cu113